add recent reading sessions hook
