+++ src/components/Verses/RecentReadingSessions.tsx
@@ -1,8 +1,6 @@
-import React, { useContext, useMemo } from 'react';
+import React, { useContext } from 'react';
 
 import useTranslation from 'next-translate/useTranslation';
-import { useSelector, shallowEqual } from 'react-redux';
-import useSWRImmutable from 'swr/immutable';
 
 import Link from '../dls/Link/Link';
 
@@ -10,44 +8,23 @@ import styles from './RecentReadingSessions.module.scss';
 import RecentReadingSessionsSkeleton from './RecentReadingSessionsSkeleton';
 
 import SurahPreview, { SurahPreviewDisplay } from '@/dls/SurahPreview/SurahPreview';
-import { selectRecentReadingSessions } from '@/redux/slices/QuranReader/readingTracker';
-import { privateFetcher } from '@/utils/auth/api';
-import { makeReadingSessionsUrl } from '@/utils/auth/apiPaths';
-import { isLoggedIn } from '@/utils/auth/login';
+import useGetRecentlyReadVerseKeys from '@/hooks/auth/useGetRecentlyReadVerseKeys';
 import { getChapterData } from '@/utils/chapter';
 import { logButtonClick } from '@/utils/eventLogger';
 import { toLocalizedNumber } from '@/utils/locale';
 import { getChapterWithStartingVerseUrl } from '@/utils/navigation';
-import { getVerseAndChapterNumbersFromKey, makeVerseKey } from '@/utils/verse';
+import { getVerseAndChapterNumbersFromKey } from '@/utils/verse';
 import DataContext from 'src/contexts/DataContext';
-import ReadingSession from 'types/ReadingSession';
 
 const RecentReadingSessions = () => {
   const { t, lang } = useTranslation('home');
   const chaptersData = useContext(DataContext);
-  const recentReadingSessions = useSelector(selectRecentReadingSessions, shallowEqual);
+  const { recentlyReadVerseKeys, isLoading } = useGetRecentlyReadVerseKeys();
   const onRecentReadingSessionClicked = () => {
     logButtonClick('homepage_recently_read_card');
   };
 
-  const { data, isValidating } = useSWRImmutable<ReadingSession[]>(
-    isLoggedIn() ? makeReadingSessionsUrl() : null,
-    privateFetcher,
-  );
-
-  const recentReadVerseKeys = useMemo(() => {
-    if (!isLoggedIn()) {
-      return Object.keys(recentReadingSessions);
-    }
-
-    if (isLoggedIn() && data?.length > 0) {
-      return data.map((item) => makeVerseKey(item.chapterNumber, item.verseNumber));
-    }
-
-    return [];
-  }, [data, recentReadingSessions]);
-
-  if (isValidating) {
+  if (isLoading) {
     return (
       <div className={styles.sessionsContainer}>
         <p className={styles.sessionsHeader}>{t('recently-read')}</p>
@@ -58,13 +35,13 @@ const RecentReadingSessions = () => {
     );
   }
 
-  if (recentReadVerseKeys.length === 0) return null;
+  if (recentlyReadVerseKeys.length === 0) return null;
 
   return (
     <div className={styles.sessionsContainer}>
       <p className={styles.sessionsHeader}>{t('recently-read')}</p>
       <div className={styles.verseLinksContainer}>
-        {recentReadVerseKeys.map((verseKey) => {
+        {recentlyReadVerseKeys.map((verseKey) => {
           const [chapterId, verseNumber] = getVerseAndChapterNumbersFromKey(verseKey);
           const surah = getChapterData(chaptersData, chapterId);
           return (
